requirements.txt
asgiref==3.8.1
Django==4.2.13
exceptiongroup==1.2.1
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.1
pytest-django==4.8.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
pillow==10.3.0
freezegun==1.5.1